chore: add chains to alert and remove toast (#3415)
* remove toast and add chains to alert

* Update Layout.tsx
